+++ lib/delayed_job.rb
@@ -1,4 +1,3 @@
-require 'active_support'
 require 'delayed/compatibility'
 require 'delayed/exceptions'
 require 'delayed/message_sending'
@@ -14,6 +13,7 @@ require 'delayed/lifecycle'
 require 'delayed/plugin'
 require 'delayed/plugins/clear_locks'
 require 'delayed/backend/base'
+require 'delayed/backend/job_preparer'
 require 'delayed/worker'
 require 'delayed/deserialization_error'
 require 'delayed/railtie' if defined?(Rails::Railtie)